chore: upgrade to 5.3.x
