--- .storybook/addons.js
@@ -1,8 +1,0 @@
-import "storybook-readme/register";
-import "@storybook/addon-knobs/register";
-import "@storybook/addon-options/register";
-import "@storybook/addon-actions/register";
-import "@storybook/addon-a11y/register";
-
-
-import "./styles.css";
