+++ .storybook/main.js
@@ -1,0 +1,10 @@
+module.exports = {
+  stories: ["../stories/index.js"],
+  addons: [
+    "@storybook/addon-knobs",
+    "storybook-readme",
+    "@storybook/addon-options",
+    "@storybook/addon-actions",
+    "@storybook/addon-a11y",
+  ],
+};
