+++ package.json
@@ -30,12 +30,12 @@
         "@babel/core": "^7.14.6",
         "@babel/preset-env": "^7.14.7",
         "@babel/preset-react": "^7.14.5",
-        "@storybook/addon-a11y": "^5.1.9",
-        "@storybook/addon-actions": "^5.1.9",
-        "@storybook/addon-info": "^5.1.9",
-        "@storybook/addon-knobs": "^5.1.9",
-        "@storybook/addon-options": "^5.1.9",
-        "@storybook/react": "^5.1.9",
+        "@storybook/addon-a11y": "^5.3.0",
+        "@storybook/addon-actions": "^5.3.0",
+        "@storybook/addon-info": "^5.3.0",
+        "@storybook/addon-knobs": "^5.3.0",
+        "@storybook/addon-options": "^5.3.0",
+        "@storybook/react": "^5.3.0",
         "@storybook/storybook-deployer": "^2.8.1",
         "babel-loader": "^8.2.2",
         "css-loader": "^0.27.3",
